+++ vod/mp4/mp4_clipper.c
@@ -1506,7 +1506,7 @@ mp4_clipper_process_moov_atom_callback(void* ctx, atom_info_t* atom_info)
 	if (media_type != MEDIA_TYPE_NONE)
 	{
 		track_index = context->track_indexes[media_type]++;
-		if ((context->parse_params.required_tracks_mask[media_type] & (1 << track_index)) == 0)
+		if ((context->parse_params.required_tracks_mask[media_type] & ((uint64_t)1 << track_index)) == 0)
 		{
 			return VOD_OK;
 		}
